completed stretch goal functionality for feature-confirm-delete

+++ server/public/client.js
@@ -83,16 +83,33 @@ function completeTask(){
 
 function deleteTask(){
     let taskId = $(this).data('id');
-    $.ajax({
-        method:'DELETE',
-        url: `/list/${taskId}`
-    })
-    .then(response => {
-        getTasks();
-    })
-    .catch(error => {
-        alert('Error deleting record', error);
-    });
+    Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+            $.ajax({
+                method:'DELETE',
+                url: `/list/${taskId}`
+            })
+            .then(response => {
+                getTasks();
+            })
+            .catch(error => {
+                alert('Error deleting record', error);
+            });
+          Swal.fire(
+            'Deleted!',
+            'Your task has been deleted.',
+            'success'
+          )
+        }
+      })
 }
 
 /*
